+++ shared/client/client.go
@@ -2,10 +2,13 @@ package client
 
 import (
 	"bytes"
+	"context"
+	b64 "encoding/base64"
 	"encoding/json"
 	"fmt"
 	"net/http"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/pkg/errors"
 )
 
@@ -13,21 +16,27 @@ type TokenProvider interface {
 	GetToken() (string, error)
 }
 
-type HappyClient struct {
-	client        http.Client
-	apiBaseUrl    string
-	clientName    string
-	clientVersion string
-	tokenProvider TokenProvider
+type AWSCredentialsProvider interface {
+	GetCredentials(context.Context) (aws.Credentials, error)
 }
 
-func NewHappyClient(clientName, clientVersion, apiBaseUrl string, tokenProvider TokenProvider) *HappyClient {
+type HappyClient struct {
+	client          http.Client
+	apiBaseUrl      string
+	clientName      string
+	clientVersion   string
+	tokenProvider   TokenProvider
+	awsCredProvider AWSCredentialsProvider
+}
+
+func NewHappyClient(clientName, clientVersion, apiBaseUrl string, tokenProvider TokenProvider, awsCredProvider AWSCredentialsProvider) *HappyClient {
 	return &HappyClient{
-		apiBaseUrl:    apiBaseUrl,
-		clientName:    clientName,
-		clientVersion: clientVersion,
-		client:        http.Client{},
-		tokenProvider: tokenProvider,
+		apiBaseUrl:      apiBaseUrl,
+		clientName:      clientName,
+		clientVersion:   clientVersion,
+		client:          http.Client{},
+		tokenProvider:   tokenProvider,
+		awsCredProvider: awsCredProvider,
 	}
 }
 
@@ -110,9 +119,17 @@ func (c *HappyClient) Do(req *http.Request) (*http.Response, error) {
 func (c *HappyClient) addAuth(req *http.Request) error {
 	token, err := c.tokenProvider.GetToken()
 	if err != nil {
-		return errors.Wrap(err, "failed to get token")
+		return errors.Wrap(err, "failed to get oidc token")
 	}
 	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", token))
+
+	creds, err := c.awsCredProvider.GetCredentials(context.Background())
+	if err != nil {
+		return errors.Wrap(err, "failed to get aws credentials")
+	}
+	req.Header.Add("x-aws-access-key-id", b64.StdEncoding.EncodeToString([]byte(creds.AccessKeyID)))
+	req.Header.Add("x-aws-secret-access-key", b64.StdEncoding.EncodeToString([]byte(creds.SecretAccessKey)))
+	req.Header.Add("x-aws-session-token", creds.SessionToken) // SessionToken is already base64 encoded
 
 	return nil
 }