+++ terraform/modules/happy-service-eks/tags.tf
@@ -1,5 +1,5 @@
 locals {
-  tags = {
+  tags = merge(var.tags, {
     happy_env          = var.deployment_stage,
     happy_stack_name   = var.stack_name,
     happy_service_name = var.routing.service_name,
@@ -7,12 +7,12 @@ locals {
     happy_image        = var.image,
     happy_service_type = var.service_type,
     happy_last_applied = timestamp(),
-  }
+  })
 
-  routing_tags = {
+  routing_tags = merge(var.tags, {
     happy_env          = var.deployment_stage,
     happy_stack_name   = var.stack_name,
     happy_region       = data.aws_region.current.name,
     happy_last_applied = timestamp(),
-  }
+  })
 }